Update grammar.md
I refined the grammar a bit:
- removed left recursion
- separated transitions to terminal symbols from transitions to other rules
- fixed some mistakes (like the variable with leading number thing)
I created and tested this grammar with ANTLR4, so it should be correct (besides the logical operation error mentioned in the file)
I'm not quiet sure if this belongs here since it is way bigger than the previous version and also a bit more oriented towards parser / lexer building. If you like I could try to shrink it down.