UnaryOp Breaking Change
This proposal add a new node type which moves the increment and decrement operators out of the unary op node and into a new node type. This simplifies parsing of the AST by removing the need to handle errenous programs (e.g. `++(a*b)`) which can no longer be represented by the AST.